make memory db backend and flatmap backend different classes and selectable at runtime
