+++ src/ripple/app/rdb/backend/MemoryDatabase.h
@@ -12,30 +12,7 @@
 #include <shared_mutex>
 #include <vector>
 
-#if RDB_CONCURRENT
-#include <boost/unordered/concurrent_flat_map.hpp>
-#endif
-
 namespace ripple {
-
-#if RDB_CONCURRENT
-struct base_uint_hasher
-{
-    using result_type = std::size_t;
-
-    result_type
-    operator()(base_uint<256> const& value) const
-    {
-        return hardened_hash<>{}(value);
-    }
-
-    result_type
-    operator()(AccountID const& value) const
-    {
-        return hardened_hash<>{}(value);
-    }
-};
-#endif
 
 class MemoryDatabase : public SQLiteDatabase
 {
@@ -43,52 +20,26 @@ private:
     struct LedgerData
     {
         LedgerInfo info;
-#if RDB_CONCURRENT
-        boost::unordered::
-            concurrent_flat_map<uint256, AccountTx, base_uint_hasher>
-                transactions;
-#else
         std::map<uint256, AccountTx> transactions;
-#endif
     };
 
     struct AccountTxData
     {
-#if RDB_CONCURRENT
-        boost::unordered::
-            concurrent_flat_map<std::pair<uint32_t, uint32_t>, AccountTx>
-                transactions;
-#else
         AccountTxs transactions;
         std::map<uint32_t, std::map<uint32_t, size_t>>
             ledgerTxMap;  // ledgerSeq -> txSeq -> index in transactions
-#endif
     };
 
     Application& app_;
     Config const& config_;
     JobQueue& jobQueue_;
 
-#if !RDB_CONCURRENT
     mutable std::shared_mutex mutex_;
-#endif
 
-#if RDB_CONCURRENT
-    boost::unordered::concurrent_flat_map<LedgerIndex, LedgerData> ledgers_;
-    boost::unordered::
-        concurrent_flat_map<uint256, LedgerIndex, base_uint_hasher>
-            ledgerHashToSeq_;
-    boost::unordered::concurrent_flat_map<uint256, AccountTx, base_uint_hasher>
-        transactionMap_;
-    boost::unordered::
-        concurrent_flat_map<AccountID, AccountTxData, base_uint_hasher>
-            accountTxMap_;
-#else
     std::map<LedgerIndex, LedgerData> ledgers_;
     std::map<uint256, LedgerIndex> ledgerHashToSeq_;
     std::map<uint256, AccountTx> transactionMap_;
     std::map<AccountID, AccountTxData> accountTxMap_;
-#endif
 
 public:
     MemoryDatabase(Application& app, Config const& config, JobQueue& jobQueue)
@@ -99,59 +50,24 @@ public:
     std::optional<LedgerIndex>
     getMinLedgerSeq() override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerIndex> minSeq;
-        ledgers_.visit_all([&minSeq](auto const& pair) {
-            if (!minSeq || pair.first < *minSeq)
-            {
-                minSeq = pair.first;
-            }
-        });
-        return minSeq;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (ledgers_.empty())
             return std::nullopt;
         return ledgers_.begin()->first;
-#endif
     }
 
     std::optional<LedgerIndex>
     getTransactionsMinLedgerSeq() override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerIndex> minSeq;
-        transactionMap_.visit_all([&minSeq](auto const& pair) {
-            LedgerIndex seq = pair.second.second->getLgrSeq();
-            if (!minSeq || seq < *minSeq)
-            {
-                minSeq = seq;
-            }
-        });
-        return minSeq;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (transactionMap_.empty())
             return std::nullopt;
         return transactionMap_.begin()->second.second->getLgrSeq();
-#endif
     }
 
     std::optional<LedgerIndex>
     getAccountTransactionsMinLedgerSeq() override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerIndex> minSeq;
-        accountTxMap_.visit_all([&minSeq](auto const& pair) {
-            pair.second.transactions.visit_all([&minSeq](auto const& tx) {
-                if (!minSeq || tx.first.first < *minSeq)
-                {
-                    minSeq = tx.first.first;
-                }
-            });
-        });
-        return minSeq;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (accountTxMap_.empty())
             return std::nullopt;
@@ -165,45 +81,19 @@ public:
         return minSeq == std::numeric_limits<LedgerIndex>::max()
             ? std::nullopt
             : std::optional<LedgerIndex>(minSeq);
-#endif
     }
 
     std::optional<LedgerIndex>
     getMaxLedgerSeq() override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerIndex> maxSeq;
-        ledgers_.visit_all([&maxSeq](auto const& pair) {
-            if (!maxSeq || pair.first > *maxSeq)
-            {
-                maxSeq = pair.first;
-            }
-        });
-        return maxSeq;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (ledgers_.empty())
             return std::nullopt;
         return ledgers_.rbegin()->first;
-#endif
     }
     void
     deleteTransactionByLedgerSeq(LedgerIndex ledgerSeq) override
     {
-#if RDB_CONCURRENT
-        ledgers_.visit(ledgerSeq, [this](auto& item) {
-            item.second.transactions.visit_all([this](auto const& txPair) {
-                transactionMap_.erase(txPair.first);
-            });
-            item.second.transactions.clear();
-        });
-
-        accountTxMap_.visit_all([ledgerSeq](auto& item) {
-            item.second.transactions.erase_if([ledgerSeq](auto const& tx) {
-                return tx.first.first == ledgerSeq;
-            });
-        });
-#else
         std::unique_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.find(ledgerSeq);
         if (it != ledgers_.end())
@@ -226,31 +116,11 @@ public:
                     }),
                 accountData.transactions.end());
         }
-#endif
     }
 
     void
     deleteBeforeLedgerSeq(LedgerIndex ledgerSeq) override
     {
-#if RDB_CONCURRENT
-        ledgers_.erase_if([this, ledgerSeq](auto const& item) {
-            if (item.first < ledgerSeq)
-            {
-                item.second.transactions.visit_all([this](auto const& txPair) {
-                    transactionMap_.erase(txPair.first);
-                });
-                ledgerHashToSeq_.erase(item.second.info.hash);
-                return true;
-            }
-            return false;
-        });
-
-        accountTxMap_.visit_all([ledgerSeq](auto& item) {
-            item.second.transactions.erase_if([ledgerSeq](auto const& tx) {
-                return tx.first.first < ledgerSeq;
-            });
-        });
-#else
         std::unique_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.begin();
         while (it != ledgers_.end() && it->first < ledgerSeq)
@@ -279,29 +149,11 @@ public:
                     }),
                 accountData.transactions.end());
         }
-#endif
     }
 
     void
     deleteTransactionsBeforeLedgerSeq(LedgerIndex ledgerSeq) override
     {
-#if RDB_CONCURRENT
-        ledgers_.visit_all([this, ledgerSeq](auto& item) {
-            if (item.first < ledgerSeq)
-            {
-                item.second.transactions.visit_all([this](auto const& txPair) {
-                    transactionMap_.erase(txPair.first);
-                });
-                item.second.transactions.clear();
-            }
-        });
-
-        accountTxMap_.visit_all([ledgerSeq](auto& item) {
-            item.second.transactions.erase_if([ledgerSeq](auto const& tx) {
-                return tx.first.first < ledgerSeq;
-            });
-        });
-#else
         std::unique_lock<std::shared_mutex> lock(mutex_);
         for (auto& [seq, ledgerData] : ledgers_)
         {
@@ -331,19 +183,11 @@ public:
                     }),
                 accountData.transactions.end());
         }
-#endif
     }
 
     void
     deleteAccountTransactionsBeforeLedgerSeq(LedgerIndex ledgerSeq) override
     {
-#if RDB_CONCURRENT
-        accountTxMap_.visit_all([ledgerSeq](auto& item) {
-            item.second.transactions.erase_if([ledgerSeq](auto const& tx) {
-                return tx.first.first < ledgerSeq;
-            });
-        });
-#else
         std::unique_lock<std::shared_mutex> lock(mutex_);
         for (auto& [_, accountData] : accountTxMap_)
         {
@@ -362,29 +206,17 @@ public:
                     }),
                 accountData.transactions.end());
         }
-#endif
     }
     std::size_t
     getTransactionCount() override
     {
-#if RDB_CONCURRENT
-        return transactionMap_.size();
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         return transactionMap_.size();
-#endif
     }
 
     std::size_t
     getAccountTransactionCount() override
     {
-#if RDB_CONCURRENT
-        std::size_t count = 0;
-        accountTxMap_.visit_all([&count](auto const& item) {
-            count += item.second.transactions.size();
-        });
-        return count;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::size_t count = 0;
         for (const auto& [_, accountData] : accountTxMap_)
@@ -392,34 +224,16 @@ public:
             count += accountData.transactions.size();
         }
         return count;
-#endif
     }
 
     CountMinMax
     getLedgerCountMinMax() override
     {
-#if RDB_CONCURRENT
-        CountMinMax result{0, 0, 0};
-        ledgers_.visit_all([&result](auto const& item) {
-            result.numberOfRows++;
-            if (result.minLedgerSequence == 0 ||
-                item.first < result.minLedgerSequence)
-            {
-                result.minLedgerSequence = item.first;
-            }
-            if (item.first > result.maxLedgerSequence)
-            {
-                result.maxLedgerSequence = item.first;
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (ledgers_.empty())
             return {0, 0, 0};
         return {
             ledgers_.size(), ledgers_.begin()->first, ledgers_.rbegin()->first};
-#endif
     }
 
     bool
@@ -427,93 +241,6 @@ public:
         std::shared_ptr<Ledger const> const& ledger,
         bool current) override
     {
-#if RDB_CONCURRENT
-        try
-        {
-            LedgerData ledgerData;
-            ledgerData.info = ledger->info();
-
-            auto aLedger = std::make_shared<AcceptedLedger>(ledger, app_);
-            for (auto const& acceptedLedgerTx : *aLedger)
-            {
-                auto const& txn = acceptedLedgerTx->getTxn();
-                auto const& meta = acceptedLedgerTx->getMeta();
-                auto const& id = txn->getTransactionID();
-
-                std::string reason;
-                auto accTx = std::make_pair(
-                    std::make_shared<ripple::Transaction>(txn, reason, app_),
-                    std::make_shared<ripple::TxMeta>(meta));
-
-                ledgerData.transactions.emplace(id, accTx);
-                transactionMap_.emplace(id, accTx);
-
-                for (auto const& account : meta.getAffectedAccounts())
-                {
-                    accountTxMap_.visit(account, [&](auto& data) {
-                        data.second.transactions.emplace(
-                            std::make_pair(
-                                ledger->info().seq,
-                                acceptedLedgerTx->getTxnSeq()),
-                            accTx);
-                    });
-                }
-            }
-
-            ledgers_.emplace(ledger->info().seq, std::move(ledgerData));
-            ledgerHashToSeq_.emplace(ledger->info().hash, ledger->info().seq);
-
-            if (current)
-            {
-                auto const cutoffSeq =
-                    ledger->info().seq > app_.config().LEDGER_HISTORY
-                    ? ledger->info().seq - app_.config().LEDGER_HISTORY
-                    : 0;
-
-                if (cutoffSeq > 0)
-                {
-                    const std::size_t BATCH_SIZE = 128;
-                    std::size_t deleted = 0;
-
-                    ledgers_.erase_if([&](auto const& item) {
-                        if (deleted >= BATCH_SIZE)
-                            return false;
-
-                        if (item.first < cutoffSeq)
-                        {
-                            item.second.transactions.visit_all(
-                                [this](auto const& txPair) {
-                                    transactionMap_.erase(txPair.first);
-                                });
-                            ledgerHashToSeq_.erase(item.second.info.hash);
-                            deleted++;
-                            return true;
-                        }
-                        return false;
-                    });
-
-                    if (deleted > 0)
-                    {
-                        accountTxMap_.visit_all([cutoffSeq](auto& item) {
-                            item.second.transactions.erase_if(
-                                [cutoffSeq](auto const& tx) {
-                                    return tx.first.first < cutoffSeq;
-                                });
-                        });
-                    }
-
-                    app_.getLedgerMaster().clearPriorLedgers(cutoffSeq);
-                }
-            }
-
-            return true;
-        }
-        catch (std::exception const&)
-        {
-            deleteTransactionByLedgerSeq(ledger->info().seq);
-            return false;
-        }
-#else
         std::unique_lock<std::shared_mutex> lock(mutex_);
         LedgerData ledgerData;
         ledgerData.info = ledger->info();
@@ -643,139 +370,69 @@ public:
         }
 
         return true;
-#endif
     }
 
     std::optional<LedgerInfo>
     getLedgerInfoByIndex(LedgerIndex ledgerSeq) override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerInfo> result;
-        ledgers_.visit(ledgerSeq, [&result](auto const& item) {
-            result = item.second.info;
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.find(ledgerSeq);
         if (it != ledgers_.end())
             return it->second.info;
         return std::nullopt;
-#endif
     }
 
     std::optional<LedgerInfo>
     getNewestLedgerInfo() override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerInfo> result;
-        ledgers_.visit_all([&result](auto const& item) {
-            if (!result || item.second.info.seq > result->seq)
-            {
-                result = item.second.info;
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         if (ledgers_.empty())
             return std::nullopt;
         return ledgers_.rbegin()->second.info;
-#endif
     }
 
     std::optional<LedgerInfo>
     getLimitedOldestLedgerInfo(LedgerIndex ledgerFirstIndex) override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerInfo> result;
-        ledgers_.visit_all([&](auto const& item) {
-            if (item.first >= ledgerFirstIndex &&
-                (!result || item.first < result->seq))
-            {
-                result = item.second.info;
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.lower_bound(ledgerFirstIndex);
         if (it != ledgers_.end())
             return it->second.info;
         return std::nullopt;
-#endif
     }
 
     std::optional<LedgerInfo>
     getLimitedNewestLedgerInfo(LedgerIndex ledgerFirstIndex) override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerInfo> result;
-        ledgers_.visit_all([&](auto const& item) {
-            if (item.first >= ledgerFirstIndex &&
-                (!result || item.first > result->seq))
-            {
-                result = item.second.info;
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.lower_bound(ledgerFirstIndex);
         if (it == ledgers_.end())
             return std::nullopt;
         return ledgers_.rbegin()->second.info;
-#endif
     }
 
     std::optional<LedgerInfo>
     getLedgerInfoByHash(uint256 const& ledgerHash) override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerInfo> result;
-        ledgerHashToSeq_.visit(ledgerHash, [this, &result](auto const& item) {
-            ledgers_.visit(item.second, [&result](auto const& item) {
-                result = item.second.info;
-            });
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgerHashToSeq_.find(ledgerHash);
         if (it != ledgerHashToSeq_.end())
             return ledgers_.at(it->second).info;
         return std::nullopt;
-#endif
     }
     uint256
     getHashByIndex(LedgerIndex ledgerIndex) override
     {
-#if RDB_CONCURRENT
-        uint256 result;
-        ledgers_.visit(ledgerIndex, [&result](auto const& item) {
-            result = item.second.info.hash;
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.find(ledgerIndex);
         if (it != ledgers_.end())
             return it->second.info.hash;
         return uint256();
-#endif
     }
 
     std::optional<LedgerHashPair>
     getHashesByIndex(LedgerIndex ledgerIndex) override
     {
-#if RDB_CONCURRENT
-        std::optional<LedgerHashPair> result;
-        ledgers_.visit(ledgerIndex, [&result](auto const& item) {
-            result = LedgerHashPair{
-                item.second.info.hash, item.second.info.parentHash};
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = ledgers_.find(ledgerIndex);
         if (it != ledgers_.end())
@@ -784,23 +441,11 @@ public:
                 it->second.info.hash, it->second.info.parentHash};
         }
         return std::nullopt;
-#endif
     }
 
     std::map<LedgerIndex, LedgerHashPair>
     getHashesByIndex(LedgerIndex minSeq, LedgerIndex maxSeq) override
     {
-#if RDB_CONCURRENT
-        std::map<LedgerIndex, LedgerHashPair> result;
-        ledgers_.visit_all([&](auto const& item) {
-            if (item.first >= minSeq && item.first <= maxSeq)
-            {
-                result[item.first] = LedgerHashPair{
-                    item.second.info.hash, item.second.info.parentHash};
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::map<LedgerIndex, LedgerHashPair> result;
         auto it = ledgers_.lower_bound(minSeq);
@@ -811,7 +456,6 @@ public:
                 it->second.info.hash, it->second.info.parentHash};
         }
         return result;
-#endif
     }
 
     std::variant<AccountTx, TxSearched>
@@ -820,23 +464,6 @@ public:
         std::optional<ClosedInterval<std::uint32_t>> const& range,
         error_code_i& ec) override
     {
-#if RDB_CONCURRENT
-        std::variant<AccountTx, TxSearched> result = TxSearched::unknown;
-        transactionMap_.visit(id, [&](auto const& item) {
-            auto const& tx = item.second;
-            if (!range ||
-                (range->lower() <= tx.second->getLgrSeq() &&
-                 tx.second->getLgrSeq() <= range->upper()))
-            {
-                result = tx;
-            }
-            else
-            {
-                result = TxSearched::all;
-            }
-        });
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = transactionMap_.find(id);
         if (it != transactionMap_.end())
@@ -863,7 +490,6 @@ public:
         }
 
         return TxSearched::unknown;
-#endif
     }
 
     bool
@@ -881,18 +507,6 @@ public:
     std::uint32_t
     getKBUsedAll() override
     {
-#if RDB_CONCURRENT
-        std::uint32_t size = sizeof(*this);
-        size += ledgers_.size() * (sizeof(LedgerIndex) + sizeof(LedgerData));
-        size +=
-            ledgerHashToSeq_.size() * (sizeof(uint256) + sizeof(LedgerIndex));
-        size += transactionMap_.size() * (sizeof(uint256) + sizeof(AccountTx));
-        accountTxMap_.visit_all([&size](auto const& item) {
-            size += sizeof(AccountID) + sizeof(AccountTxData);
-            size += item.second.transactions.size() * sizeof(AccountTx);
-        });
-        return size / 1024;  // Convert to KB
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::uint32_t size = sizeof(*this);
         size += ledgers_.size() * (sizeof(LedgerIndex) + sizeof(LedgerData));
@@ -910,40 +524,22 @@ public:
             }
         }
         return size / 1024;
-#endif
     }
 
     std::uint32_t
     getKBUsedLedger() override
     {
-#if RDB_CONCURRENT
-        std::uint32_t size =
-            ledgers_.size() * (sizeof(LedgerIndex) + sizeof(LedgerData));
-        size +=
-            ledgerHashToSeq_.size() * (sizeof(uint256) + sizeof(LedgerIndex));
-        return size / 1024;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::uint32_t size = 0;
         size += ledgers_.size() * (sizeof(LedgerIndex) + sizeof(LedgerData));
         size +=
             ledgerHashToSeq_.size() * (sizeof(uint256) + sizeof(LedgerIndex));
         return size / 1024;
-#endif
     }
 
     std::uint32_t
     getKBUsedTransaction() override
     {
-#if RDB_CONCURRENT
-        std::uint32_t size =
-            transactionMap_.size() * (sizeof(uint256) + sizeof(AccountTx));
-        accountTxMap_.visit_all([&size](auto const& item) {
-            size += sizeof(AccountID) + sizeof(AccountTxData);
-            size += item.second.transactions.size() * sizeof(AccountTx);
-        });
-        return size / 1024;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::uint32_t size = 0;
         size += transactionMap_.size() * (sizeof(uint256) + sizeof(AccountTx));
@@ -958,7 +554,6 @@ public:
             }
         }
         return size / 1024;
-#endif
     }
 
     void
@@ -975,20 +570,6 @@ public:
 
     ~MemoryDatabase()
     {
-#if RDB_CONCURRENT
-        // Concurrent maps need visit_all
-        accountTxMap_.visit_all(
-            [](auto& pair) { pair.second.transactions.clear(); });
-        accountTxMap_.clear();
-
-        transactionMap_.clear();
-
-        ledgers_.visit_all(
-            [](auto& pair) { pair.second.transactions.clear(); });
-        ledgers_.clear();
-
-        ledgerHashToSeq_.clear();
-#else
         // Regular maps can use standard clear
         accountTxMap_.clear();
         transactionMap_.clear();
@@ -998,30 +579,11 @@ public:
         }
         ledgers_.clear();
         ledgerHashToSeq_.clear();
-#endif
     }
 
     std::vector<std::shared_ptr<Transaction>>
     getTxHistory(LedgerIndex startIndex) override
     {
-#if RDB_CONCURRENT
-        std::vector<std::shared_ptr<Transaction>> result;
-        transactionMap_.visit_all([&](auto const& item) {
-            if (item.second.second->getLgrSeq() >= startIndex)
-            {
-                result.push_back(item.second.first);
-            }
-        });
-        std::sort(
-            result.begin(), result.end(), [](auto const& a, auto const& b) {
-                return a->getLedger() > b->getLedger();
-            });
-        if (result.size() > 20)
-        {
-            result.resize(20);
-        }
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         std::vector<std::shared_ptr<Transaction>> result;
         auto it = ledgers_.lower_bound(startIndex);
@@ -1037,7 +599,6 @@ public:
             ++it;
         }
         return result;
-#endif
     }
     // Helper function to handle limits
     template <typename Container>
@@ -1053,24 +614,6 @@ public:
     AccountTxs
     getOldestAccountTxs(AccountTxOptions const& options) override
     {
-#if RDB_CONCURRENT
-        AccountTxs result;
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    result.push_back(tx.second);
-                }
-            });
-        });
-        std::sort(
-            result.begin(), result.end(), [](auto const& a, auto const& b) {
-                return a.second->getLgrSeq() < b.second->getLgrSeq();
-            });
-        applyLimit(result, options.limit, options.bUnlimited);
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1100,30 +643,11 @@ public:
         }
 
         return result;
-#endif
     }
 
     AccountTxs
     getNewestAccountTxs(AccountTxOptions const& options) override
     {
-#if RDB_CONCURRENT
-        AccountTxs result;
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    result.push_back(tx.second);
-                }
-            });
-        });
-        std::sort(
-            result.begin(), result.end(), [](auto const& a, auto const& b) {
-                return a.second->getLgrSeq() > b.second->getLgrSeq();
-            });
-        applyLimit(result, options.limit, options.bUnlimited);
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1156,37 +680,11 @@ public:
         }
 
         return result;
-#endif
     }
 
     MetaTxsList
     getOldestAccountTxsB(AccountTxOptions const& options) override
     {
-#if RDB_CONCURRENT
-        MetaTxsList result;
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    result.emplace_back(
-                        tx.second.first->getSTransaction()
-                            ->getSerializer()
-                            .peekData(),
-                        tx.second.second->getAsObject()
-                            .getSerializer()
-                            .peekData(),
-                        tx.first.first);
-                }
-            });
-        });
-        std::sort(
-            result.begin(), result.end(), [](auto const& a, auto const& b) {
-                return std::get<2>(a) < std::get<2>(b);
-            });
-        applyLimit(result, options.limit, options.bUnlimited);
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1220,37 +718,11 @@ public:
         }
 
         return result;
-#endif
     }
 
     MetaTxsList
     getNewestAccountTxsB(AccountTxOptions const& options) override
     {
-#if RDB_CONCURRENT
-        MetaTxsList result;
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    result.emplace_back(
-                        tx.second.first->getSTransaction()
-                            ->getSerializer()
-                            .peekData(),
-                        tx.second.second->getAsObject()
-                            .getSerializer()
-                            .peekData(),
-                        tx.first.first);
-                }
-            });
-        });
-        std::sort(
-            result.begin(), result.end(), [](auto const& a, auto const& b) {
-                return std::get<2>(a) > std::get<2>(b);
-            });
-        applyLimit(result, options.limit, options.bUnlimited);
-        return result;
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1288,56 +760,10 @@ public:
         }
 
         return result;
-#endif
     }
     std::pair<AccountTxs, std::optional<AccountTxMarker>>
     oldestAccountTxPage(AccountTxPageOptions const& options) override
     {
-#if RDB_CONCURRENT
-        AccountTxs result;
-        std::optional<AccountTxMarker> marker;
-
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            std::vector<std::pair<std::pair<uint32_t, uint32_t>, AccountTx>>
-                txs;
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    txs.emplace_back(tx);
-                }
-            });
-
-            std::sort(txs.begin(), txs.end(), [](auto const& a, auto const& b) {
-                return a.first < b.first;
-            });
-
-            auto it = txs.begin();
-            if (options.marker)
-            {
-                it = std::find_if(txs.begin(), txs.end(), [&](auto const& tx) {
-                    return tx.first.first == options.marker->ledgerSeq &&
-                        tx.first.second == options.marker->txnSeq;
-                });
-                if (it != txs.end())
-                    ++it;
-            }
-
-            for (; it != txs.end() &&
-                 (options.limit == 0 || result.size() < options.limit);
-                 ++it)
-            {
-                result.push_back(it->second);
-            }
-
-            if (it != txs.end())
-            {
-                marker = AccountTxMarker{it->first.first, it->first.second};
-            }
-        });
-
-        return {result, marker};
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1377,57 +803,11 @@ public:
         }
 
         return {result, marker};
-#endif
     }
 
     std::pair<AccountTxs, std::optional<AccountTxMarker>>
     newestAccountTxPage(AccountTxPageOptions const& options) override
     {
-#if RDB_CONCURRENT
-        AccountTxs result;
-        std::optional<AccountTxMarker> marker;
-
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            std::vector<std::pair<std::pair<uint32_t, uint32_t>, AccountTx>>
-                txs;
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    txs.emplace_back(tx);
-                }
-            });
-
-            std::sort(txs.begin(), txs.end(), [](auto const& a, auto const& b) {
-                return a.first > b.first;
-            });
-
-            auto it = txs.begin();
-            if (options.marker)
-            {
-                it = std::find_if(txs.begin(), txs.end(), [&](auto const& tx) {
-                    return tx.first.first == options.marker->ledgerSeq &&
-                        tx.first.second == options.marker->txnSeq;
-                });
-                if (it != txs.end())
-                    ++it;
-            }
-
-            for (; it != txs.end() &&
-                 (options.limit == 0 || result.size() < options.limit);
-                 ++it)
-            {
-                result.push_back(it->second);
-            }
-
-            if (it != txs.end())
-            {
-                marker = AccountTxMarker{it->first.first, it->first.second};
-            }
-        });
-
-        return {result, marker};
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1471,59 +851,11 @@ public:
         }
 
         return {result, marker};
-#endif
     }
 
     std::pair<MetaTxsList, std::optional<AccountTxMarker>>
     oldestAccountTxPageB(AccountTxPageOptions const& options) override
     {
-#if RDB_CONCURRENT
-        MetaTxsList result;
-        std::optional<AccountTxMarker> marker;
-
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            std::vector<std::tuple<uint32_t, uint32_t, AccountTx>> txs;
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    txs.emplace_back(
-                        tx.first.first, tx.first.second, tx.second);
-                }
-            });
-
-            std::sort(txs.begin(), txs.end());
-
-            auto it = txs.begin();
-            if (options.marker)
-            {
-                it = std::find_if(txs.begin(), txs.end(), [&](auto const& tx) {
-                    return std::get<0>(tx) == options.marker->ledgerSeq &&
-                        std::get<1>(tx) == options.marker->txnSeq;
-                });
-                if (it != txs.end())
-                    ++it;
-            }
-
-            for (; it != txs.end() &&
-                 (options.limit == 0 || result.size() < options.limit);
-                 ++it)
-            {
-                const auto& [_, __, tx] = *it;
-                result.emplace_back(
-                    tx.first->getSTransaction()->getSerializer().peekData(),
-                    tx.second->getAsObject().getSerializer().peekData(),
-                    std::get<0>(*it));
-            }
-
-            if (it != txs.end())
-            {
-                marker = AccountTxMarker{std::get<0>(*it), std::get<1>(*it)};
-            }
-        });
-
-        return {result, marker};
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1567,59 +899,11 @@ public:
         }
 
         return {result, marker};
-#endif
     }
 
     std::pair<MetaTxsList, std::optional<AccountTxMarker>>
     newestAccountTxPageB(AccountTxPageOptions const& options) override
     {
-#if RDB_CONCURRENT
-        MetaTxsList result;
-        std::optional<AccountTxMarker> marker;
-
-        accountTxMap_.visit(options.account, [&](auto const& item) {
-            std::vector<std::tuple<uint32_t, uint32_t, AccountTx>> txs;
-            item.second.transactions.visit_all([&](auto const& tx) {
-                if (tx.first.first >= options.minLedger &&
-                    tx.first.first <= options.maxLedger)
-                {
-                    txs.emplace_back(
-                        tx.first.first, tx.first.second, tx.second);
-                }
-            });
-
-            std::sort(txs.begin(), txs.end(), std::greater<>());
-
-            auto it = txs.begin();
-            if (options.marker)
-            {
-                it = std::find_if(txs.begin(), txs.end(), [&](auto const& tx) {
-                    return std::get<0>(tx) == options.marker->ledgerSeq &&
-                        std::get<1>(tx) == options.marker->txnSeq;
-                });
-                if (it != txs.end())
-                    ++it;
-            }
-
-            for (; it != txs.end() &&
-                 (options.limit == 0 || result.size() < options.limit);
-                 ++it)
-            {
-                const auto& [_, __, tx] = *it;
-                result.emplace_back(
-                    tx.first->getSTransaction()->getSerializer().peekData(),
-                    tx.second->getAsObject().getSerializer().peekData(),
-                    std::get<0>(*it));
-            }
-
-            if (it != txs.end())
-            {
-                marker = AccountTxMarker{std::get<0>(*it), std::get<1>(*it)};
-            }
-        });
-
-        return {result, marker};
-#else
         std::shared_lock<std::shared_mutex> lock(mutex_);
         auto it = accountTxMap_.find(options.account);
         if (it == accountTxMap_.end())
@@ -1668,7 +952,6 @@ public:
         }
 
         return {result, marker};
-#endif
     }
 };
